Rename package from idtoken to openid.
