+++ src/test/kotlin/org/wfanet/measurement/common/crypto/tink/BUILD.bazel
@@ -34,6 +34,6 @@ kt_jvm_test(
         "//imports/kotlin/kotlin/test",
         "//src/main/kotlin/org/wfanet/measurement/common/crypto/tink",
         "//src/main/kotlin/org/wfanet/measurement/common/identity",
-        "//src/main/kotlin/org/wfanet/measurement/common/idtoken:request_uri",
+        "//src/main/kotlin/org/wfanet/measurement/common/openid:request_uri",
     ],
 )
